requirement.txt
torch>=1.6
torchvision
numpy
pandas
tqdm
lmdb
opencv-python
pillow
tensorboardx
wandb
opencv-python
